+++ pkg/cache/v2/lru.go
@@ -6,30 +6,30 @@ import (
 )
 
 // CacheItem is an item in the LRU cache
-type CacheItem[V any] struct {
-	key   string
+type CacheItem[K comparable, V any] struct {
+	key   K
 	value V
 }
 
 // LRUCache is a generic type Least Recently Used (LRU) cache
-type LRUCache[V any] struct {
+type LRUCache[K comparable, V any] struct {
 	maxItems int
-	items    map[string]*list.Element
+	items    map[K]*list.Element
 	list     *list.List
 	mutex    sync.Mutex
 }
 
 // NewLRUCache returns a new instance of the LRUCache with specified capacity
-func NewLRUCache[V any](maxItems int) *LRUCache[V] {
-	return &LRUCache[V]{
+func NewLRUCache[K comparable, V any](maxItems int) *LRUCache[K, V] {
+	return &LRUCache[K, V]{
 		maxItems: maxItems,
-		items:    make(map[string]*list.Element),
+		items:    make(map[K]*list.Element),
 		list:     list.New(),
 	}
 }
 
 // Len returns the number of items in the cache
-func (c *LRUCache[V]) Len() int {
+func (c *LRUCache[K, V]) Len() int {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
@@ -37,13 +37,13 @@ func (c *LRUCache[V]) Len() int {
 }
 
 // Get retrieves an item from the cache by key. Returns the value and true if the item exists, otherwise false
-func (c *LRUCache[V]) Get(key string) (V, bool) {
+func (c *LRUCache[K, V]) Get(key K) (V, bool) {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
 	if elem, ok := c.items[key]; ok {
 		c.list.MoveToFront(elem)
-		return elem.Value.(*CacheItem[V]).value, true
+		return elem.Value.(*CacheItem[K, V]).value, true
 	}
 
 	var result V
@@ -52,22 +52,22 @@ func (c *LRUCache[V]) Get(key string) (V, bool) {
 
 // Put adds an item to the cache. If the item already exists, update its value and move it to the front of the list.
 // If the cache is full, remove the least recently used item before adding the new item.
-func (c *LRUCache[V]) Put(key string, value V) {
+func (c *LRUCache[K, V]) Put(key K, value V) {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
 	if elem, ok := c.items[key]; ok {
 		c.list.MoveToFront(elem)
-		elem.Value.(*CacheItem[V]).value = value
+		elem.Value.(*CacheItem[K, V]).value = value
 		return
 	}
 
-	item := &CacheItem[V]{key: key, value: value}
+	item := &CacheItem[K, V]{key: key, value: value}
 
 	if c.list.Len() >= c.maxItems {
 		// remove the least recently used item
 		elem := c.list.Back()
-		item := elem.Value.(*CacheItem[V])
+		item := elem.Value.(*CacheItem[K, V])
 		delete(c.items, item.key)
 		c.list.Remove(elem)
 	}
@@ -77,7 +77,7 @@ func (c *LRUCache[V]) Put(key string, value V) {
 }
 
 // Delete removes an item from the cache by key. Returns true if the item exists and removed, otherwise false.
-func (c *LRUCache[V]) Delete(key string) bool {
+func (c *LRUCache[K, V]) Delete(key K) bool {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
@@ -91,10 +91,10 @@ func (c *LRUCache[V]) Delete(key string) bool {
 }
 
 // Clear removes all items from the cache
-func (c *LRUCache[V]) Clear() {
+func (c *LRUCache[K, V]) Clear() {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
 	c.list.Init()
-	c.items = make(map[string]*list.Element)
+	c.items = make(map[K]*list.Element)
 }